requirements.txt
ccxt>=4.0.0
redis>=5.0.0
pandas>=2.0.0
pandas_ta
streamlit>=1.30.0
pyyaml>=6.0
python-dotenv>=1.0.0
matplotlib>=3.0.0
scikit-learn>=1.3.0
joblib>=1.3.0
